CDEC-416: Sin bin Pos.Infra.Util.JsonLog.Events
